refactor: change place for function yup
